backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
faiss-cpu
numpy
pandas
tqdm
python-dotenv
openai
sentence-transformers
torch
transformers
